Disabling js interop in Typescript build of backend-core and worker, fixing build issues.
