+++ packages/backend-core/src/cache/index.ts
@@ -2,3 +2,4 @@ export * as generic from "./generic"
 export * as user from "./user"
 export * as app from "./appMetadata"
 export * as writethrough from "./writethrough"
+export * from "./generic"
