+++ packages/backend-core/src/middleware/auditLog.ts
@@ -1,6 +1,6 @@
 import { BBContext } from "@budibase/types"
 
-export = async (ctx: BBContext, next: any) => {
+export = async (ctx: BBContext | any, next: any) => {
   // Placeholder for audit log middleware
   return next()
 }
