+++ packages/worker/src/environment.ts
@@ -85,7 +85,7 @@ if (!environment.APPS_URL) {
 }
 
 // clean up any environment variable edge cases
-for (let [key, value] of Object.entries(env)) {
+for (let [key, value] of Object.entries(environment)) {
   // handle the edge case of "0" to disable an environment variable
   if (value === "0") {
     // @ts-ignore
